+++ asset-platform-idm-plus/src/main/java/com/asset/bean/UserRole.java
@@ -4,6 +4,7 @@ import com.baomidou.mybatisplus.annotation.IdType;
 import com.baomidou.mybatisplus.annotation.TableField;
 import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
+import io.swagger.models.auth.In;
 import lombok.Data;
 
 import java.io.Serializable;
@@ -19,7 +20,7 @@ public class UserRole implements Serializable {
 
     private String uid;
     @TableField("role_id")
-    private Long roleId;
+    private Integer roleId;
     @TableField("created_time")
     private Date createdTime;
