+++ asset-platform-idm-plus/src/main/java/com/asset/mapper/UserMapper.java
@@ -34,7 +34,7 @@ public interface UserMapper extends BaseMapper<User> {
 
     int updateByPrimaryKey(User record);
 
-    List<User> getUsersByRole(@Param("roleId") Long roleId);
+    List<User> getUsersByRole(@Param("roleId") Integer roleId);
 
     List<User> getUsersByScene(String sceneId);
